KERN-1646 works now with mimetypes for a correct file extension

+++ scripts/preview_processor.rb
@@ -43,6 +43,19 @@ def resize_and_write_file filename, filename_output, max_width, max_height
   content
 end
 
+def process_as_image? extension
+  ['.png', '.jpg', '.gif', '.psd'].include? extension
+end
+
+def determine_file_extension_with_mime_type mimetype
+  fe = `grep #{mimetype} ../mime.types`.gsub(mimetype, '').strip
+  if fe == '' || fe.nil?
+    ''
+  else
+    ".#{fe}"
+  end
+end
+
 res = @s.execute_get(@s.url_for("var/search/needsprocessing.json"))
 raise "Failed to retrieve list to process [#{res.code}]" unless res.code == '200'
 process = JSON.parse(res.body)
@@ -71,15 +84,17 @@ Dir["*"].each do |id|
     meta_file = @s.execute_get @s.url_for("p/#{id}.json")
     next unless meta_file.code == '200' # skip.
     meta = JSON.parse meta_file.body
-    extension = File.extname meta['sakai:pooled-content-file-name']
-    raise "File extension is nil" if extension.nil?
 
     # making a local copy of the file.
+    mimeType = meta['_mimeType']
+    extension = determine_file_extension_with_mime_type mimeType
     filename = id + extension
+    puts "with filename: #{filename}"
+
     content_file = @s.execute_get @s.url_for("p/#{id}")
     File.open(filename, 'wb') { |f| f.write content_file.body }
 
-    if ['.png', '.jpg', '.jpeg', '.gif', '.psd'].include? extension
+    if process_as_image? extension
       # Images don't need a preview so we make a big and small thumbnail instead.
 
       page_count = 1
@@ -98,7 +113,7 @@ Dir["*"].each do |id|
       puts "Uploaded thumb to curl #{@s.url_for("p/#{id}.page1-small.jpg")}"
 
       ## Cleaning crew.
-      FileUtils.rm DOCS_DIR + "/#{filename_thumb}"
+      FileUtils.rm DOCS_DIR + "/#{filename_thumb}" unless DEBUG
     else
       # Generating image previews of te document.
       Docsplit.extract_images filename, :size => '700x', :format => :jpg
@@ -136,12 +151,12 @@ Dir["*"].each do |id|
       end
 
       # Cleaning crew.
-      FileUtils.remove_dir PREVS_DIR + "/#{id}"
+      FileUtils.remove_dir PREVS_DIR + "/#{id}" unless DEBUG
 
     end
 
     # cleaning crew.
-    FileUtils.rm DOCS_DIR + "/#{filename}"
+    FileUtils.rm DOCS_DIR + "/#{filename}" unless DEBUG
 
     # passing on the page_count.
     @s.execute_post @s.url_for("p/#{id}"), {"sakai:pagecount" => page_count}
@@ -155,5 +170,5 @@ Dir["*"].each do |id|
   end
 end
 
-FileUtils.remove_dir PREVS_DIR
-FileUtils.remove_dir DOCS_DIR
+FileUtils.remove_dir PREVS_DIR unless DEBUG
+FileUtils.remove_dir DOCS_DIR unless DEBUG
